Back/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.114.2
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.7
passlib
bcrypt==4.0.1
PyJWT
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
openpyxl==3.1.5
orjson==3.10.7
outcome==1.3.0.post0
pycparser==2.22
pydantic==2.9.1
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.3
Pygments==2.18.0
PyPDF2==3.0.1
pyperclip==1.9.0
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
selenium==4.23.1
Send2Trash==1.8.3
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
psycopg2==2.9.9
soupsieve==2.6
starlette==0.38.5
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.6
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.0.1
wsproto==1.2.0
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.30.0
attrs==24.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.1
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.114.2
fastapi-cli==0.0.5
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
openpyxl==3.1.5
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.1
passlib==1.7.4
pillow==11.2.1
pip==24.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
Pygments==2.18.0
PyJWT==2.10.1
PyPDF2==3.0.1
pyperclip==1.9.0
PySocks==1.7.1
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
rsa==4.9.1
selenium==4.23.1
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.43
starlette==0.38.5
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
unicorn==2.1.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.0.1
wsproto==1.2.0
SQLAlchemy
asyncpg
psycopg2-binary
python-jose[cryptography]
Pillow
rapidfuzz==3.9.3
